Resolve compiler warnings

+++ daemon/main/Scheduler.cpp
@@ -2,6 +2,7 @@
  *  This file is part of nzbget. See <https://nzbget.com>.
  *
  *  Copyright (C) 2008-2019 Andrey Prygunkov <hugbug@users.sourceforge.net>
+ *  Copyright (C) 2024 Denis <denis@nzbget.com>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -65,7 +66,7 @@ void Scheduler::ScheduleNextWork()
 	sched.tm_sec = 0;
 	time_t nextMinute = Util::Timegm(&sched);
 
-	m_serviceInterval = nextMinute - curTime;
+	m_serviceInterval = static_cast<int>(nextMinute - curTime);
 }
 
 void Scheduler::ServiceWork()
